test_requirements.txt
pep8
pytest
pytest-cov
mock
requests
redis
gevent
sure
websockets